refactor: extract findClassPositions scanner to own file
Moves the text scanner into src/plugin/find-class-positions.ts with
21 unit tests covering all edge cases: CSS, SCSS, Less, Sass indented,
Stylus, block/line comments, strings, escaped quotes, &.nested,
compound selectors, URL paths, grid templates, Less mixins, pseudo-
selectors, CSS-escaped class names, and empty input.
